remove limiting in api calls remove people model from shared folder
add crawler
