+++ data_fetcher/src/people/services/people_pipeline.py
@@ -7,7 +7,7 @@ from typing import Dict, List, Optional, Callable
 from tqdm import tqdm
 from shared.src.core.logging import get_main_fetcher_logger
 from .people_service import PeopleService
-from shared.src.models.people_model import Person
+from api.src.v1.people.models.people_model import Person
 
 from .people_data_normalizer import PeopleDataNormalizer
 from .people_enum_mapper import PeopleEnumMapper
